New name for SupportingDocumentComponent
This renames `DocumentComponent` to `SupportingDocumentComponent` in
preparation for it to deal with supporting documents (ActiveStorage
attachments) instead of `Document` records.

This doesn't actually change the kind of object passed to the component
yet, just a separate PR to make the upcoming changes less confusing.
