--- app/components/document_component/document_component.html.slim
@@ -1,3 +1,0 @@
-= tag.div(class: classes, **html_attributes) do
-  = open_in_new_tab_link_to document.name, document_path(document), class: "document-component__link"
-  = document_size
