+++ app/components/jobseekers/vacancy_details_component/vacancy_details_component.html.slim
@@ -38,4 +38,4 @@ section#job-details class="govuk-!-margin-bottom-5"
       h3.govuk-heading-l.section-heading = t("jobs.supporting_documents")
       p.govuk-body = t("jobs.supporting_documents_accessibility")
       .grey-border-box--thin
-        = render DocumentComponent.with_collection(@vacancy.documents)
+        = render SupportingDocumentComponent.with_collection(@vacancy.documents)
