+++ app/components/supporting_document_component/supporting_document_component.html.slim
@@ -1,0 +1,3 @@
+= tag.div(class: classes, **html_attributes) do
+  = open_in_new_tab_link_to supporting_document.name, document_path(supporting_document), class: "supporting-document-component__link"
+  = document_size
