+++ app/components/supporting_document_component/supporting_document_component.scss
@@ -1,6 +1,6 @@
 @import 'base_component';
 
-.document-component {
+.supporting-document-component {
   margin-bottom: govuk-spacing(2);
 
   &:last-child {
